requirements.txt
azure-core==1.26.3
azure-storage-blob==12.15.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
idna==3.4
isodate==0.6.1
numpy==1.24.2
pandas==1.5.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14
